requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
Django==3.1.4
django-filter==2.4.0
djangorestframework==3.12.2
gunicorn==20.1.0
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.6.0
pylint==2.6.0
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
whitenoise==6.2.0
wrapt==1.12.1
